+++ agate/table/print_html.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+# pylint: disable=W0212
 
 import sys
 
@@ -32,15 +33,15 @@ def print_html(self, max_rows=20, max_columns=6, output=sys.stdout, max_column_w
         By default it will use the system's setting.
     """
     if max_rows is None:
-        max_rows = len(self.rows)
+        max_rows = len(self._rows)
 
     if max_columns is None:
-        max_columns = len(self.columns)
+        max_columns = len(self._columns)
 
-    rows_truncated = max_rows < len(self.rows)
-    columns_truncated = max_columns < len(self.column_names)
+    rows_truncated = max_rows < len(self._rows)
+    columns_truncated = max_columns < len(self._column_names)
 
-    column_names = list(self.column_names[:max_columns])
+    column_names = list(self._column_names[:max_columns])
 
     if columns_truncated:
         column_names.append(utils.ELLIPSIS)
@@ -49,7 +50,7 @@ def print_html(self, max_rows=20, max_columns=6, output=sys.stdout, max_column_w
     formatted_data = []
 
     # Determine correct number of decimal places for each Number column
-    for i, c in enumerate(self.columns):
+    for i, c in enumerate(self._columns):
         if i >= max_columns:
             break
 
@@ -60,7 +61,7 @@ def print_html(self, max_rows=20, max_columns=6, output=sys.stdout, max_column_w
             number_formatters.append(None)
 
     # Format data
-    for i, row in enumerate(self.rows):
+    for i, row in enumerate(self._rows):
         if i >= max_rows:
             break
 
@@ -101,7 +102,7 @@ def print_html(self, max_rows=20, max_columns=6, output=sys.stdout, max_column_w
 
         for j, d in enumerate(formatted_row):
             # Text is left-justified, all other values are right-justified
-            if isinstance(self.column_types[j], Text):
+            if isinstance(self._column_types[j], Text):
                 write('<td style="text-align: left;">%s</td>' % d)
             else:
                 write('<td style="text-align: right;">%s</td>' % d)