+++ agate/table/print_structure.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+# pylint: disable=W0212
 
 import sys
 
@@ -14,8 +15,8 @@ def print_structure(self, output=sys.stdout):
     """
     from agate.table import Table
 
-    name_column = [n for n in self.column_names]
-    type_column = [t.__class__.__name__ for t in self.column_types]
+    name_column = [n for n in self._column_names]
+    type_column = [t.__class__.__name__ for t in self._column_types]
     rows = zip(name_column, type_column)
     column_names = ['column', 'data_type']
     text = Text()