+++ agate/table/to_json.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+# pylint: disable=W0212
 
 import codecs
 from collections import OrderedDict