+++ tests/test_utils.py
@@ -10,6 +10,9 @@ try:
 except ImportError:
     import unittest
 
+import sys
+import warnings
+
 from agate.data_types import Text
 from agate.mapped_sequence import MappedSequence
 from agate.table import Table
@@ -34,6 +37,11 @@ class TryPatchShadow(object):
 
 
 class TestMonkeyPatching(unittest.TestCase):
+    def setUp(self):
+        # Workaround for https://bugs.python.org/issue4180
+        if sys.version_info[0] == 2 or sys.version_info[:2] == (3, 3):
+            warnings.simplefilter('always')
+
     def test_monkeypatch(self):
         before_table = Table([], ['foo'], [Text()])
 
@@ -52,6 +60,13 @@ class TestMonkeyPatching(unittest.TestCase):
         self.assertEqual(before_table.test(5), 5)
         self.assertEqual(after_table.test(5), 5)
         self.assertEqual(Table.testcls(5), 5)
+
+    def test_monkeypatch_deprecated(self):
+        with warnings.catch_warnings():
+            warnings.simplefilter('error')
+
+            with self.assertRaises(DeprecationWarning):
+                Table.monkeypatch(TryPatch)
 
     def test_monkeypatch_shadow(self):
         before_table = Table([['blah'], ], ['foo'], [Text()])